requirements.txt
oslo-config
oslo-log
openstacksdk
